% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_net.R
\name{GLMpath}
\alias{GLMpath}
\alias{summary.GLMpath}
\alias{print.GLMpath}
\alias{GLMpath-class}
\title{Methods for 'GLMpath' objects.}
\usage{
\method{summary}{GLMpath}(object, ...)

\method{print}{GLMpath}(x, ...)
}
\arguments{
\item{object}{\code{GLMpath} object}

\item{...}{not used}

\item{x}{\code{GLMpath} object}
}
\description{
Methods for the result of \code{\link{GLMnetwork}}
(\code{GLMpath} object)
}
\seealso{
\code{\link{GLMnetwork}}
}
\author{
{Alyssa Imbert, \email{alyssa.imbert@inra.fr}

Nathalie Villa-Vialaneix, \email{nathalie.villa-vialaneix@inra.fr}}
}
