% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donors.R
\name{HDImputed}
\alias{HDImputed}
\alias{summary.HDimputed}
\alias{summary.HDImputed}
\alias{print.HDImputed}
\title{Methods for 'HDImputed' objects.}
\usage{
\method{summary}{HDimputed}(object, ...)

\method{print}{HDImputed}(x, ...)
}
\arguments{
\item{object}{\code{HDImputed} object}

\item{...}{not used}

\item{x}{\code{HDImputed} object}
}
\description{
Methods for the result of \code{\link{imputeHD}}
(\code{HDImputed} object)
}
\seealso{
\code{\link{imputeHD}}
}
\author{
{Alyssa Imbert, \email{alyssa.imbert@inra.fr}

Nathalie Villa-Vialaneix, \email{nathalie.villa-vialaneix@inra.fr}}
}
