% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung}
\alias{lung}
\title{RNA-seq expression from lung tissue (GTEx).}
\format{a data frame with 221 rows and 100 variables (genes). Rownames are
identifiers for individuals.}
\source{
The raw data were download from \url{https://gtexportal.org/}.
The TMM normalisation of RNA-seq expression was performed with the R package
\code{edgeR}.
}
\description{
This data set is a small subset of the full data set from GTEx.  It contains
RNA-seq expressions measured from lung tissue. The RNA-seq expressions have
been normalized with the TMM method.
}
\author{
Alyssa Imbert <alyssa.imbert@inra.fr>
}
