% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_net.R
\name{stars}
\alias{stars}
\alias{summary.stars}
\alias{print.stars}
\alias{plot.stars}
\title{Methods for 'stars' objects.}
\usage{
\method{summary}{stars}(object, ...)

\method{print}{stars}(x, ...)

\method{plot}{stars}(x, ...)
}
\arguments{
\item{object}{\code{stars} object}

\item{...}{not used}

\item{x}{\code{stars} object}
}
\description{
Methods for the result of \code{\link{stabilitySelection}}
(\code{stars} object)
}
\seealso{
\code{\link{stabilitySelection}}
}
\author{
{Alyssa Imbert, \email{alyssa.imbert@inra.fr}

Nathalie Villa-Vialaneix, \email{nathalie.villa-vialaneix@inra.fr}}
}
