% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thyroid}
\alias{thyroid}
\title{RNA-seq expression from thyroid tissue (GTEx).}
\format{
a data frame with 221 rows and 50 variables (genes). Rownames are
identifiers for individuals.
}
\source{
The raw data were downloaded from \url{https://gtexportal.org/}. The
TMM normalisation of RNA-seq expression was performed with the R package
\code{edgeR}.
}
\description{
This data set is a small subset of the full data set from GTEx.  It contains
RNA-seq expressions measured from thyroid tissue. The RNA-seq expressions
have been normalized with the TMM method.
}
\author{
Alyssa Imbert <alyssa.imbert@gmail.com>
}
